# (trivial) trace_parser.py utility: hide runtime events from stats by default